Adding support for large arrays using X231 format extension
- EclIO classes have been updated to support arrays with size beyond numeric limits for integers
- Output of arrays having 2^31 or more elements is facilitated by using a pair of headers
- The first header is of type X231, second should have the actual type
- The large array format extension is documented in the Eclipse File format manual

+++ opm/io/eclipse/EclFile.hpp
@@ -52,7 +52,7 @@ public:
       char_array.clear();
     }
 
-    using EclEntry = std::tuple<std::string, eclArrType, int>;
+    using EclEntry = std::tuple<std::string, eclArrType, long int>;
     std::vector<EclEntry> getList() const;
 
     template <typename T>
@@ -78,7 +78,7 @@ protected:
 
     std::vector<std::string> array_name;
     std::vector<eclArrType> array_type;
-    std::vector<int> array_size;
+    std::vector<long int> array_size;
 
     std::vector<unsigned long int> ifStreamPos;
